feat: adiciona verificação de autenticação e permissões no hook useDeleteGame para garantir que apenas usuários autorizados possam deletar jogos

+++ src/hooks/useDeleteGame.ts
@@ -1,10 +1,12 @@
 import { db } from "@/config/firebase";
 import { TIMEOUT_TO_REMOVE_TOAST } from "@/constants";
-import { ToastType } from "@/types";
-import { deleteDoc, doc } from "firebase/firestore";
+import { useAuth } from "@/contexts/AuthContext";
+import { DatabaseSchemaType, ToastType } from "@/types";
+import { deleteDoc, doc, getDoc } from "firebase/firestore";
 import { useState } from "react";
 
 const useDeleteGame = (id: string, path: string) => {
+  const { user } = useAuth();
   const [isDeleted, setIsDeleted] = useState(false);
   const [systemMessage, setSystemMessage] = useState<ToastType>({
     message: "",
@@ -12,26 +14,42 @@ const useDeleteGame = (id: string, path: string) => {
   });
 
   const handleDelete = async () => {
-    setSystemMessage({
-      message: "",
-      variant: null,
-    });
+    setSystemMessage({ message: "", variant: null });
+
+    if (!user) {
+      setSystemMessage({
+        message: "Você precisa estar logado",
+        variant: "danger",
+      });
+      setIsDeleted(false);
+      return;
+    }
 
     try {
       const deleteRef = doc(db, path, id);
+      const docSnap = await getDoc(deleteRef);
+
+      if (!docSnap.exists()) {
+        throw new Error("Jogo não encontrado.");
+      }
+
+      const gameData = docSnap.data() as DatabaseSchemaType;
+      if (gameData.user_id !== user.uid) {
+        throw new Error("Você não tem permissão para deletar este jogo.");
+      }
+
       await deleteDoc(deleteRef);
 
-      setIsDeleted(true);
       setSystemMessage({ message: "O jogo foi deletado!", variant: "success" });
+      setIsDeleted(true);
+      console.log("chegou até aqui");
 
       setTimeout(() => {
-        setSystemMessage({
-          message: "",
-          variant: null,
-        });
+        console.log("timeou finalizou");
+        setSystemMessage({ message: "", variant: null });
       }, TIMEOUT_TO_REMOVE_TOAST);
     } catch (error) {
-      console.log(error);
+      console.error(error);
       setIsDeleted(false);
       setSystemMessage({
         message: "Algo deu errado ao deletar o jogo, tente novamente.",
